+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './RootEditor';
+export * from './RootReadonlyEditor';
 export * from './Editor.types';