+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # root-editor
 
+## 0.1.6-alpha.2
+
+### Patch Changes
+
+- feat: add readonly editor features
+
+## 0.1.6-alpha.1
+
+### Patch Changes
+
+- refactor: change the button type to button
+
+## 0.1.6-alpha.0
+
+### Patch Changes
+
+- chore: add editor types
+
 ## 0.1.5
 
 ### Patch Changes